pyproject.toml
[project]
name = "auto-congratulate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.2.1",
    "telethon>=1.42.0",
]
